requirements.txt
fastapi
uvicorn
python-dotenv
motor
stripe
python-jose[cryptography]
